Add verification when deleting a client. Closes #72

+++ src/app/clients/clients.component.ts
@@ -1,10 +1,11 @@
 // clients.component
 
 import { Component, OnInit } from '@angular/core';
-import { MatSnackBar, MatDialog, MatChipInputEvent } from '@angular/material';
+import { MatSnackBar, MatDialog, MatDialogRef, MAT_DIALOG_DATA, MatChipInputEvent } from '@angular/material';
 import { PublicFunctions } from '../shared/shared';
 import { ClientsService } from './clients.service';
 import { OpenRegisterClientComponent } from '../auth/open-register-client/open-register-client.component';
+import { VerifyDeleteComponent } from './verify-delete/verify-delete.component';
 import { ENTER, COMMA } from '@angular/cdk/keycodes';
 
 const COLORS = ['#EF5350', '#C62828', '#EC407A', '#AB47BC', '#7E57C2', '#5C6BC0',
@@ -33,7 +34,10 @@ export class ClientsComponent implements OnInit {
   * @param registerDialog - The service of the register dialog.
   * @param clientsService - The service of the clients.
   */
-  constructor(private snackBar: MatSnackBar, private registerDialog: MatDialog, private clientsService: ClientsService) { }
+  constructor(private snackBar: MatSnackBar,
+              private registerDialog: MatDialog,
+              private verifyDeleteDialog: MatDialog,
+              private clientsService: ClientsService) {}
 
   /**
    * When the component initialized, check if the account team is logged in.
@@ -228,15 +232,24 @@ export class ClientsComponent implements OnInit {
    * @param client
    */
   removeClient(client): void {
-    this.clientsService.removeClient(client.clientId).subscribe((data) => {
-      this.clients.forEach((currClient, index) => {
-        if (currClient.clientId === client.clientId) {
-          this.clients.splice(index, 1);
+    const dialogRef = this.registerDialog.open(VerifyDeleteComponent, {
+      width: '420px',
+      height: '220px',
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.clientsService.removeClient(client.clientId).subscribe((data) => {
           this.snackBar.open('Client was removed successfuly', '', {
             duration: 2000
           });
-        }
-      });
+          this.clients.forEach((currClient, index) => {
+            if (currClient.clientId === client.clientId) {
+              this.clients.splice(index, 1);
+            }
+          });
+        });
+      }
     });
   }
